+++ src/main/java/com/actiontech/dble/route/parser/druid/impl/DruidUpdateParser.java
@@ -9,7 +9,6 @@ import com.actiontech.dble.config.model.ERTable;
 import com.actiontech.dble.config.model.SchemaConfig;
 import com.actiontech.dble.config.model.TableConfig;
 import com.actiontech.dble.config.privileges.ShardingPrivileges;
-import com.actiontech.dble.plan.common.ptr.StringPtr;
 import com.actiontech.dble.route.RouteResultset;
 import com.actiontech.dble.route.parser.druid.ServerSchemaStatVisitor;
 import com.actiontech.dble.route.parser.util.Pair;
@@ -19,6 +18,7 @@ import com.actiontech.dble.server.util.SchemaUtil;
 import com.actiontech.dble.server.util.SchemaUtil.SchemaInfo;
 import com.actiontech.dble.util.StringUtil;
 import com.alibaba.druid.sql.ast.SQLExpr;
+import com.alibaba.druid.sql.ast.SQLObject;
 import com.alibaba.druid.sql.ast.SQLStatement;
 import com.alibaba.druid.sql.ast.expr.*;
 import com.alibaba.druid.sql.ast.statement.*;
@@ -26,17 +26,17 @@ import com.alibaba.druid.sql.dialect.mysql.ast.statement.MySqlUpdateStatement;
 
 import java.sql.SQLException;
 import java.sql.SQLNonTransientException;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 
 /**
  * see http://dev.mysql.com/doc/refman/5.7/en/update.html
  *
  * @author huqing.yan
  */
-public class DruidUpdateParser extends DefaultDruidParser {
+public class DruidUpdateParser extends DruidModifyParser {
+
+    protected static final String MODIFY_SQL_NOT_SUPPORT_MESSAGE = "This `Complex Update Syntax` is not supported!";
+
     @Override
     public SchemaConfig visitorParse(SchemaConfig schema, RouteResultset rrs, SQLStatement stmt, ServerSchemaStatVisitor visitor, ServerConnection sc, boolean isExplain)
             throws SQLException {
@@ -44,14 +44,35 @@ public class DruidUpdateParser extends DefaultDruidParser {
         SQLTableSource tableSource = update.getTableSource();
         String schemaName = schema == null ? null : schema.getName();
         if (tableSource instanceof SQLJoinTableSource) {
-            StringPtr noShardingNode = new StringPtr(null);
-            Set<String> schemas = new HashSet<>();
-            if (!SchemaUtil.isNoSharding(sc, (SQLJoinTableSource) tableSource, stmt, stmt, schemaName, schemas, noShardingNode)) {
-                String msg = "UPDATE query with multiple tables is not supported, sql:" + stmt;
-                throw new SQLNonTransientException(msg);
-            } else {
-                return routeToNoSharding(schema, rrs, schemas, noShardingNode);
+            super.visitorParse(schema, rrs, stmt, visitor, sc, isExplain);
+            if (visitor.getSubQueryList().size() > 0) {
+                throw new SQLNonTransientException(MODIFY_SQL_NOT_SUPPORT_MESSAGE);
             }
+
+            List<SchemaInfo> schemaInfos = checkPrivilegeForModifyTable(sc, schemaName, stmt, visitor.getMotifyTableSourceList());
+
+            boolean isAllGlobal = true;
+            for (SchemaInfo schemaInfo : schemaInfos) {
+                TableConfig tc = schemaInfo.getSchemaConfig().getTables().get(schemaInfo.getTable());
+                if (tc == null || !tc.isGlobalTable()) {
+                    isAllGlobal = false;
+                    break;
+                } else if (tc.getShardingNodes().size() > 1) {
+                    throw new SQLNonTransientException(MODIFY_SQL_NOT_SUPPORT_MESSAGE);
+                }
+            }
+
+            Collection<String> routeShardingNodes;
+            if (isAllGlobal) {
+                routeShardingNodes = checkForMultiNodeGlobal(schemaInfos);
+            } else {
+                //try to route to single Node for each table
+                routeShardingNodes = checkForSingleNodeTable(rrs);
+            }
+
+            RouterUtil.routeToMultiNode(false, rrs, routeShardingNodes, true);
+            rrs.setFinishedRoute(true);
+            return schema;
         } else {
             SchemaInfo schemaInfo = SchemaUtil.getSchemaInfo(sc.getUser(), schemaName, (SQLExprTableSource) tableSource);
             if (!ShardingPrivileges.checkPrivilege(sc.getUserConfig(), schemaInfo.getSchema(), schemaInfo.getTable(), ShardingPrivileges.CheckType.UPDATE)) {
@@ -61,27 +82,15 @@ public class DruidUpdateParser extends DefaultDruidParser {
             schema = schemaInfo.getSchemaConfig();
             rrs.setStatement(RouterUtil.removeSchema(rrs.getStatement(), schemaInfo.getSchema()));
             super.visitorParse(schema, rrs, stmt, visitor, sc, isExplain);
-            if (visitor.getSubQueryList().size() > 0) {
-                StringPtr noShardingNode = new StringPtr(null);
-                Set<String> schemas = new HashSet<>();
-                if (!SchemaUtil.isNoSharding(sc, tableSource, stmt, stmt, schemaInfo.getSchema(), schemas, noShardingNode)) {
-                    String msg = "UPDATE query with sub-query  is not supported, sql:" + stmt;
-                    throw new SQLNonTransientException(msg);
-                } else {
-                    return routeToNoSharding(schema, rrs, schemas, noShardingNode);
-                }
-            }
 
             String tableName = schemaInfo.getTable();
-            String noShardingNode = RouterUtil.isNoSharding(schema, tableName);
-            if (noShardingNode != null) {
-                RouterUtil.routeToSingleNode(rrs, noShardingNode);
-                rrs.setFinishedRoute(true);
-                return schema;
-            }
             TableConfig tc = schema.getTables().get(tableName);
             checkTableExists(tc, schema.getName(), tableName, ShardingPrivileges.CheckType.UPDATE);
 
+            if (visitor.getFirstClassSubQueryList().size() > 0) {
+                routeForModifySubQueryList(rrs, tc, visitor, schema);
+                return schema;
+            }
 
             if (tc.isGlobalTable()) {
                 RouterUtil.routeToMultiNode(false, rrs, tc.getShardingNodes(), tc.isGlobalTable());
@@ -109,6 +118,7 @@ public class DruidUpdateParser extends DefaultDruidParser {
         }
         return schema;
     }
+
 
     private static boolean columnInExpr(SQLExpr sqlExpr, String colName) throws SQLNonTransientException {
         String column;
@@ -272,4 +282,17 @@ public class DruidUpdateParser extends DefaultDruidParser {
         ERTable key = new ERTable(schema.getName(), tableName, column);
         return map.containsKey(key);
     }
+
+    @Override
+    int tryGetShardingColIndex(SchemaInfo schemaInfo, SQLStatement stmt, String partitionColumn) throws SQLNonTransientException {
+        return 0;
+    }
+
+    @Override
+    SQLSelect acceptVisitor(SQLObject stmt, ServerSchemaStatVisitor visitor) {
+        stmt.accept(visitor);
+        return (SQLSelect) stmt;
+    }
+
+
 }
